capstone-backend/requirements.txt
fastapi
earthengine-api
pandas
numpy
scikit-learn
requests
python-dotenv
uvicorn
streamlit
matplotlib
seaborn
motor
python-dotenv
apscheduler
python-multipart